requirement.txt
colorama==0.4.3
openai==0.20.0
python-dotenv==0.20.0